+++ lib/caching/index.ts
@@ -1,0 +1,5 @@
+'use strict';
+
+import {memoize} from './memoize';
+
+export {memoize};